requirements.txt
requests
psycopg2-binary
pandas
streamlit
plotly
dbt-postgres

statsmodels

prophet
